# api-spring/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: FinanKal Finance API
  version: 0.0.1
  description: REST API Gateway for FinanKal Ledger Engine
  contact:
    name: FinanKal Team
  license:
    name: MIT

servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.finankal.com
    description: Production Server

paths:
  /api/accounts/{id}:
    get:
      summary: Get Account Summary
      description: Retrieves a summary of an account including balance and account type
      operationId: getAccount
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          description: Account UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '400':
          description: Invalid account ID
        '404':
          description: Account not found
        '500':
          description: Server error

  /api/accounts/{id}/balance:
    get:
      summary: Get Account Balance
      description: Retrieves the current balance of an account
      operationId: getBalance
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          description: Account UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account balance retrieved successfully
          content:
            application/json:
              schema:
                type: number
                format: decimal
                example: 1000.50
        '400':
          description: Invalid account ID
        '404':
          description: Account not found
        '500':
          description: Server error

  /api/transactions:
    post:
      summary: Create Transaction
      description: Creates a new transaction with multiple ledger entries
      operationId: createTransaction
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequestDto'
      responses:
        '200':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDto'
        '400':
          description: Invalid request payload
        '500':
          description: Server error

  /api/users/{userId}/accounts:
    get:
      summary: Get User Accounts
      operationId: getUserAccounts
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountDto'

  /api/ledger-entries/{accountId}:
    get:
      summary: Get Ledger Entries
      description: Retrieves all ledger entries for a specific account
      operationId: getLedgerEntries
      tags:
        - Ledger Entries
      parameters:
        - name: accountId
          in: path
          description: Account UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ledger entries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerEntryDto'
        '400':
          description: Invalid account ID
        '404':
          description: Account not found
        '500':
          description: Server error

  /api/users/{id}/total-credit:
    get:
      summary: Get User Total Credit
      description: Retrieves the total credit amount for a user across all accounts
      operationId: getUserTotalCredit
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User total credit retrieved successfully
          content:
            application/json:
              schema:
                type: number
                format: decimal
                example: 1500.00
        '400':
          description: Invalid user ID
        '404':
          description: User not found
        '500':
          description: Server error

  /api/users/{id}/total-debit:
    get:
      summary: Get User Total Debit
      description: Retrieves the total debit amount for a user across all accounts
      operationId: getUserTotalDebit
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User total debit retrieved successfully
          content:
            application/json:
              schema:
                type: number
                format: decimal
                example: 1200.00
        '400':
          description: Invalid user ID
        '404':
          description: User not found
        '500':
          description: Server error

  /api/users/{id}/net-worth:
    get:
      summary: Get User Net Worth
      description: Retrieves the net worth for a user across all accounts
      operationId: getUserNetWorth
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User net worth retrieved successfully
          content:
            application/json:
              schema:
                type: number
                format: decimal
                example: 300.00

        '400':
          description: Invalid user ID

        '404':
          description: User not found

        '500':
          description: Server error

  /api/credit-cards:
    post:
      summary: Create Credit Card
      description: Creates a new credit card for an account. Requires either accountId or accountName.
      operationId: createCreditCard
      tags:
        - Credit Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCreditCardRequestDto'
      responses:
        '200':
          description: Credit card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCreditCardResponseDto'
        '400':
          description: Invalid request - either accountId or accountName must be provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Account lookup by name not yet implemented. Please provide accountId.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/credit-cards/transactions:
    post:
      summary: Record Purchase
      description: Records a purchase transaction on a credit card
      operationId: recordCreditCardTransaction
      tags:
        - Credit Cards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordCreditCardTransactionRequestDto'
      responses:
        '200':
          description: Purchase recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: string
                    format: uuid
        '400':
          description: Invalid request payload
        '404':
          description: Card not found
        '500':
          description: Server error

  /api/credit-cards/payment:
    post:
      summary: Pay Statement
      description: Records a payment against a credit card statement
      operationId: payCreditCardStatement
      tags:
        - Statements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayCreditCardStatementRequestDto'
      responses:
        '200':
          description: Payment recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionId:
                    type: string
                    format: uuid
        '400':
          description: Invalid request payload
        '404':
          description: Statement or card not found
        '500':
          description: Server error

  /api/credit-cards/statements/{statementId}:
    get:
      summary: Get Statement
      description: Retrieves a credit card statement
      operationId: getCreditCardStatement
      tags:
        - Statements
      parameters:
        - name: statementId
          in: path
          description: Statement UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statement retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardStatementDto'
        '400':
          description: Invalid statement ID
        '404':
          description: Statement not found
        '500':
          description: Server error

components:
  schemas:
    AccountDto:
      type: object
      description: Account information with balance
      properties:
        id:
          type: string
          format: uuid
          description: Account unique identifier
          example: "11111111-1111-1111-1111-111111111111"
        name:
          type: string
          description: Account name
          example: "Checking Account"
        type:
          type: string
          description: Account type
          example: "ASSET"
        balance:
          type: number
          format: decimal
          description: Current account balance
          example: 5000.00
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2026-01-15T10:30:00"

    CreateTransactionRequestDto:
      type: object
      required:
        - description
        - entries
      description: Request body for creating a transaction
      properties:
        description:
          type: string
          description: Transaction description
          example: "Payment for services"
        entries:
          type: array
          minItems: 2
          description: List of ledger entries for the transaction
          items:
            $ref: '#/components/schemas/CreateLedgerEntryDto'

    CreateLedgerEntryDto:
      type: object
      required:
        - accountId
        - amount
        - type
      description: Ledger entry for creating a transaction
      properties:
        accountId:
          type: string
          format: uuid
          description: Account UUID
          example: "11111111-1111-1111-1111-111111111111"
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 100.00
        type:
          type: string
          enum: [ DEBIT, CREDIT ]
          description: Entry type
          example: "DEBIT"

    TransactionDto:
      type: object
      description: Transaction information with entries
      properties:
        id:
          type: string
          format: uuid
          description: Transaction unique identifier
          example: "b8d4c031-70de-4ef5-9e83-cc12253951bc"
        description:
          type: string
          description: Transaction description
          example: "Payment for services"
        entries:
          type: array
          description: List of ledger entries
          items:
            $ref: '#/components/schemas/LedgerEntryDto'

    LedgerEntryDto:
      type: object
      description: Ledger entry details
      properties:
        accountId:
          type: string
          format: uuid
          description: Account UUID
          example: "11111111-1111-1111-1111-111111111111"
        amount:
          type: number
          format: decimal
          description: Entry amount
          example: 100.00
        type:
          type: string
          enum: [ DEBIT, CREDIT ]
          description: Entry type
          example: "DEBIT"
        description:
          type: string
          nullable: true
          description: Entry description from transaction
          example: "Payment for services"
        transactionId:
          type: string
          format: uuid
          nullable: true
          description: Associated transaction ID
          example: "b8d4c031-70de-4ef5-9e83-cc12253951bc"
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: Entry creation timestamp
          example: "2026-04-03T18:30:00"

    BalanceDto:
      type: object
      description: Balance information
      properties:
        balance:
          type: number
          format: decimal
          description: Balance amount
          example: 1000.50

    UserTotalDto:
      type: object
      description: User total amounts across all accounts
      properties:
        userId:
          type: string
          format: uuid
          description: User unique identifier
        totalCredit:
          type: number
          format: decimal
          description: Total credit amount
          example: 1500.00
        totalDebit:
          type: number
          format: decimal
          description: Total debit amount
          example: 1200.00
        totalBalance:
          type: number
          format: decimal
          description: Total balance
          example: 300.00

    CreateCreditCardRequestDto:
      type: object
      description: Request body for creating a credit card. Either accountId or accountName must be provided (accountId is recommended).
      properties:
        accountId:
          type: string
          format: uuid
          nullable: true
          description: Account UUID (required if accountName is not provided)
          example: "11111111-1111-1111-1111-111111111111"
        accountName:
          type: string
          nullable: true
          description: Account name (required if accountId is not provided, not yet implemented)
          example: "John Doe Credit Card"
        creditLimit:
          type: number
          format: decimal
          description: Credit limit amount
          example: 50000.00
        billingDay:
          type: integer
          description: Day of month for billing cycle
          example: 15
        paymentDueDays:
          type: integer
          description: Number of days until payment is due
          example: 21
      required:
        - creditLimit
        - billingDay
        - paymentDueDays

    CreateCreditCardResponseDto:
      type: object
      description: Response for credit card creation
      properties:
        creditCardId:
          type: string
          format: uuid
          description: Credit card unique identifier
          example: "11111111-1111-1111-1111-111111111111"

    RecordCreditCardTransactionRequestDto:
      type: object
      required:
        - cardId
        - amount
        - purchaseDate
      description: Request body for recording a credit card transaction
      properties:
        cardId:
          type: string
          format: uuid
          description: Credit card UUID
          example: "11111111-1111-1111-1111-111111111111"
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 1500.00
        description:
          type: string
          description: Transaction description
          example: "Coffee"
        purchaseDate:
          type: string
          format: date-time
          description: Purchase timestamp
          example: "2026-08-05T10:00:00Z"

    PayCreditCardStatementRequestDto:
      type: object
      required:
        - statementId
        - cardId
        - paymentAccountId
        - amount
      description: Request body for paying a credit card statement
      properties:
        statementId:
          type: string
          format: uuid
          description: Statement UUID
          example: "11111111-1111-1111-1111-111111111111"
        cardId:
          type: string
          format: uuid
          description: Credit card UUID
          example: "11111111-1111-1111-1111-111111111111"
        paymentAccountId:
          type: string
          format: uuid
          description: Account to pay from
          example: "11111111-1111-1111-1111-111111111111"
        amount:
          type: number
          format: decimal
          description: Payment amount
          example: 5000.00
        description:
          type: string
          description: Payment description
          example: "August payment"

    CreditCardStatementDto:
      type: object
      description: Credit card statement information
      properties:
        statementId:
          type: string
          format: uuid
          description: Statement unique identifier
          example: "11111111-1111-1111-1111-111111111111"
        creditCardId:
          type: string
          format: uuid
          description: Credit card unique identifier
          example: "11111111-1111-1111-1111-111111111111"
        startDate:
          type: string
          format: date
          description: Statement period start date
          example: "2026-08-01"
        endDate:
          type: string
          format: date
          description: Statement period end date
          example: "2026-08-31"
        statementDate:
          type: string
          format: date
          description: Statement generation date
          example: "2026-09-01"
        dueDate:
          type: string
          format: date
          description: Payment due date
          example: "2026-09-21"
        totalAmount:
          type: number
          format: decimal
          description: Total statement amount
          example: 12000.00
        status:
          type: string
          description: Statement status
          example: "OPEN"

    ErrorResponse:
      type: object
      description: Error response
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

tags:
  - name: Accounts
    description: Account management endpoints
  - name: Transactions
    description: Transaction creation endpoints
  - name: Ledger Entries
    description: Ledger entry retrieval endpoints
  - name: Users
    description: User-related endpoints
  - name: Credit Cards
    description: Credit card management endpoints
  - name: Statements
    description: Credit card statement endpoints